+++ build.sh
@@ -22,6 +22,7 @@ make_one(){
 make_one 'UAVTALK' &&
 make_one 'MAVLINK' &&
 make_one 'MWII' &&
+make_one 'NMEA' &&
 make_one 'LTM'